# Savestates: Track dirty pages for the purposes of speeding up sequential savestates