--- a/Source/Core/Common/MemArenaWin.cpp
+++ b/Source/Core/Common/MemArenaWin.cpp
@@ -34,7 +34,10 @@ using PMapViewOfFile3 = PVOID(WINAPI*)(HANDLE FileMapping, HANDLE Process, PVOID
 
 using PUnmapViewOfFileEx = BOOL(WINAPI*)(PVOID BaseAddress, ULONG UnmapFlags);
 
+using PVirtualProtectEx = BOOL(WINAPI*)(HANDLE Process, LPVOID lpAddress, SIZE_T dwSize, DWORD flNewProtect, PDWORD* lpflOldProtect);
+
 using PIsApiSetImplemented = BOOL(APIENTRY*)(PCSTR Contract);
+
 
 namespace Common
 {
@@ -78,11 +81,14 @@ static bool InitWindowsMemoryFunctions(WindowsMemoryFunctions* functions)
       functions->m_api_ms_win_core_memory_l1_1_6_handle.GetSymbolAddress("MapViewOfFile3FromApp");
   void* const address_UnmapViewOfFileEx =
       functions->m_kernel32_handle.GetSymbolAddress("UnmapViewOfFileEx");
+  void* const address_VirtualProtectEx =
+      functions->m_kernel32_handle.GetSymbolAddress("VirtualProtectEx");
   if (address_VirtualAlloc2 && address_MapViewOfFile3 && address_UnmapViewOfFileEx)
   {
     functions->m_address_VirtualAlloc2 = address_VirtualAlloc2;
     functions->m_address_MapViewOfFile3 = address_MapViewOfFile3;
     functions->m_address_UnmapViewOfFileEx = address_UnmapViewOfFileEx;
+    functions->m_address_VirtualProtectEx = address_VirtualProtectEx;
     return true;
   }
 
@@ -207,6 +213,13 @@ void MemArena::ReleaseMemoryRegion()
     m_reserved_region = nullptr;
     m_regions.clear();
   }
+}
+
+bool MemArena::WriteProtectMemoryRegion(void* data, size_t size)
+{
+  PDWORD lpflOldProtect = 0;
+  return static_cast<PVirtualProtectEx>(m_memory_functions.m_address_VirtualProtectEx)(
+      nullptr, data, size, FILE_MAP_READ, &lpflOldProtect);
 }
 
 WindowsMemoryRegion* MemArena::EnsureSplitRegionForMapping(void* start_address, size_t size)