An attempt to solve the dirty page tracking problem referenced [here](https://github.com/dolphin-emu/dolphin/pull/10873).