@JaredWhiteOne 2024-07-03
> What kind of feedback are you seeking on this PR at this stage? Are things working correctly right now?

Mostly unsure if the direction is correct, or if I should be directing efforts elsewhere. Yes, it does seem to work (at least on Windows) but I'd also like to get this tested more thoroughly and on other platforms once I implement them.