@AdmiralCurtiss 2024-07-03
The logic in the savestating code looks wrong to me. How are you recovering the position of the memory on load?