@JaredWhiteOne 2024-07-03
> The logic in the savestating code looks wrong to me. How are you recovering the position of the memory on load?

Yes, that part of the logic is quite rough and definitely wouldn't work past the first savestate in its current state; effectively my approach to solve the memory latency issues presented by needing to take sequential savestates every frame is to treat each state as a delta, where if you need to load savestate 3 after taking savestate 5 you must load savestates in reverse order from 5 -> 4 -> 3. This reduces the amount of information you need to store between savestates, or, in the case of rollback, frames.