@JaredWhiteOne 2024-07-03
> Have you measured what kind of performance improvement you get from this? Before thinking about making it robust, it would be useful to know how much there is to gain from this. For all I know, maybe something other than savestating the memory is taking most of the savestating time.

I'm seeing significant improvements in a personal project utilizing a slightly different dirty tracking method. This is a fair ask, and I will benchmark each "stating step" to make sure my theory is correct. If it's not, I will close this draft.