@JaredWhiteOne 2024-07-03
After some initial benchmarking:

**On master**
| Game  | Save Time | Load Time |
| ------------- | ------------- | ------------- |
| Super Smash Bros. Melee (On FoD, 1v1) | ~11.5 ms  | ~67.5 ms |
| Super Smash Bros. Brawl (On New Pork City, 1v1) | ~25.5 ms | ~100.5 ms |

Here are my (relevant) specs:
AMD Ryzen 9 5950X
NVIDIA GeForce GTX 1060 6GB
32 GB DDR4-3603 / PC4-28800 DDR4 SDRAM UDIMM

What's interesting is that most expensive action for saving a savestate in both games was stating within the Hardware namespace, with about ~18 ms spent stating Hardware and ~8 ms of that being the memory state in Brawl, and ~8 ms in Melee with ~5-6 ms of it being just memory. Conversely, by far the most expensive action in loading state is the PowerPC namespace, with ~55 ms in Melee and ~60 ms in Brawl.

I'd therefore argue that the most gains for saving state would indeed be with the memory portion, but with loading state it would be by far more effective to focus on the PPC state functionality.