@JosJuice 2024-07-04
At the current stage, the code doesn't look like it will even reliably function, so I don't think that's something we really have to think about at this stage. This isn't going to get merged in its current state anyway. Long-term, I guess it would make sense to have some kind of parameter to the savestating code to indicate whether you want a regular savestate or a delta savestate, maybe?